chore: moved pure calc functions under utils
