+++ src/chart_types/partition_chart/layout/utils/circline_geometry.ts
@@ -24,8 +24,8 @@ import {
   Distance,
   PointObject,
   RingSectorConstruction,
-} from './types/geometry_types';
-import { TAU } from './utils/constants';
+} from '../types/geometry_types';
+import { TAU } from './constants';
 
 function euclideanDistance({ x: x1, y: y1 }: PointObject, { x: x2, y: y2 }: PointObject): Distance {
   return Math.sqrt(Math.pow(x1 - x2, 2) + Math.pow(y1 - y2, 2));
